perf(unhead)!: composable `resolveTags()`
