+++ packages/unhead/src/unhead.ts
@@ -6,14 +6,10 @@ import type {
   HeadHooks,
   HeadPlugin,
   HeadPluginInput,
-  HeadTag,
   ResolvableHead,
   Unhead,
 } from './types'
 import { createHooks } from 'hookable'
-import { isMetaArrayDupeKey, sortTags, tagWeight, UsesMergeStrategy, ValidHeadTags } from './utils'
-import { dedupeKey } from './utils/dedupe'
-import { normalizeEntryToTags } from './utils/normalize'
 
 function registerPlugin(head: Unhead<any>, p: HeadPluginInput) {
   const plugin = (typeof p === 'function' ? p(head) : p)
@@ -47,9 +43,10 @@ export function createUnhead<T = ResolvableHead>(resolvedOptions: CreateHeadOpti
 
   const entries: Map<number, HeadEntry<T>> = new Map()
   const plugins: Map<string, HeadPlugin> = new Map()
-  const normalizeQueue = new Set<number>()
+  const _normalizeQueue = new Set<number>()
   const head: Unhead<T> = {
     _entryCount: 1, // 0 is reserved for internal use
+    _normalizeQueue,
     plugins,
     dirty: false,
     resolvedOptions,
@@ -68,7 +65,7 @@ export function createUnhead<T = ResolvableHead>(resolvedOptions: CreateHeadOpti
       const _: ActiveHeadEntry<T> = {
         _poll(rm = false) {
           head.dirty = true
-          !rm && normalizeQueue.add(_i)
+          !rm && _normalizeQueue.add(_i)
           hooks.callHook('entries:updated', head)
         },
         dispose() {
@@ -89,141 +86,10 @@ export function createUnhead<T = ResolvableHead>(resolvedOptions: CreateHeadOpti
       _.patch(input)
       return _
     },
-    resolveTags() {
-      const ctx: { tagMap: Map<string, HeadTag>, tags: HeadTag[], entries: HeadEntry<T>[] } = {
-        tagMap: new Map(),
-        tags: [],
-        entries: [...head.entries.values()],
-      }
-      // Call hooks synchronously
-      hooks.callHook('entries:resolve', ctx)
-      while (normalizeQueue.size) {
-        const i = normalizeQueue.values().next().value!
-        normalizeQueue.delete(i)
-        const e = entries.get(i)
-        if (e) {
-          const normalizeCtx = {
-            tags: normalizeEntryToTags(e.input, resolvedOptions.propResolvers || [])
-              .map(t => Object.assign(t, e.options)),
-            entry: e,
-          }
-          hooks.callHook('entries:normalize', normalizeCtx)
-          e._tags = normalizeCtx.tags.map((t, i) => {
-            t._w = tagWeight(head, t)
-            t._p = (e._i << 10) + i
-            t._d = dedupeKey(t)
-            return t
-          })
-        }
-      }
-      let hasFlatMeta = false
-      ctx.entries
-        .flatMap(e => (e._tags || []).map(t => ({ ...t, props: { ...t.props } })))
-        .sort(sortTags)
-        .reduce((acc, next) => {
-          const k = String(next._d || next._p)
-          if (!acc.has(k))
-            return acc.set(k, next)
-
-          const prev = acc.get(k)!
-          const strategy = next?.tagDuplicateStrategy || (UsesMergeStrategy.has(next.tag) ? 'merge' : null)
-            || (next.key && next.key === prev.key ? 'merge' : null)
-
-          if (strategy === 'merge') {
-            const newProps = { ...prev.props }
-            Object.entries(next.props).forEach(([p, v]) =>
-            // @ts-expect-error untyped
-              newProps[p] = p === 'style'
-              // @ts-expect-error untyped
-                ? new Map([...(prev.props.style || new Map()), ...v])
-                : p === 'class' ? new Set([...(prev.props.class || new Set()), ...v]) : v)
-            acc.set(k, { ...next, props: newProps })
-          }
-          else if ((next._p! >> 10) === (prev._p! >> 10) && next.tag === 'meta' && isMetaArrayDupeKey(k)) {
-            acc.set(k, Object.assign([...(Array.isArray(prev) ? prev : [prev]), next], next))
-            hasFlatMeta = true
-          }
-          // @ts-expect-error untyped
-          else if (next._w === prev._w ? next._p! > prev._p! : next?._w < prev?._w) {
-            acc.set(k, next)
-          }
-          return acc
-        }, ctx.tagMap)
-
-      const title = ctx.tagMap.get('title')
-      const titleTemplate = ctx.tagMap.get('titleTemplate')
-      head._title = title?.textContent
-
-      if (titleTemplate) {
-        const titleTemplateFn = titleTemplate?.textContent
-        head._titleTemplate = titleTemplateFn
-        if (titleTemplateFn) {
-          // @ts-expect-error todo
-          let newTitle = (typeof titleTemplateFn === 'function' ? titleTemplateFn(title?.textContent) : titleTemplateFn)
-          if (typeof newTitle === 'string' && !head.plugins.has('template-params')) {
-            newTitle = newTitle.replace('%s', title?.textContent || '')
-          }
-          if (title) {
-            newTitle === null
-              ? ctx.tagMap.delete('title')
-              : ctx.tagMap.set('title', { ...title, textContent: newTitle })
-          }
-          else {
-            // convert title template to a title
-            titleTemplate.tag = 'title'
-            titleTemplate.textContent = newTitle
-          }
-        }
-      }
-      // merge _tags into one map
-      ctx.tags = Array.from(ctx.tagMap!.values())
-      if (hasFlatMeta) {
-        ctx.tags = ctx.tags.flat().sort(sortTags)
-      }
-
-      hooks.callHook('tags:beforeResolve', ctx)
-      hooks.callHook('tags:resolve', ctx)
-      // post-processing mainly for XSS prevention
-      hooks.callHook('tags:afterResolve', ctx)
-      const finalTags: HeadTag[] = []
-      for (const t of ctx.tags) {
-        const { innerHTML, tag, props } = t
-        if (!ValidHeadTags.has(tag)) {
-          continue
-        }
-        // avoid rendering empty tags
-        if (Object.keys(props).length === 0 && !t.innerHTML && !t.textContent) {
-          continue
-        }
-        if (tag === 'meta' && !props.content && !props['http-equiv'] && !props.charset) {
-          continue
-        }
-        // final XSS
-        if (tag === 'script' && innerHTML) {
-          if (String(props.type).endsWith('json')) {
-            const v = typeof innerHTML === 'string' ? innerHTML : JSON.stringify(innerHTML)
-            // ensure </script> tags get encoded, this is only for JSON, it will break HTML if used
-            t.innerHTML = v.replace(/</g, '\\u003C')
-          }
-          else if (typeof innerHTML === 'string') {
-            // make sure the tag isn't being ended
-            t.innerHTML = innerHTML.replace(new RegExp(`</${tag}`, 'g'), `<\\/${tag}`)
-          }
-          // the dedupe key may have become invalid
-          t._d = dedupeKey(t)
-        }
-        finalTags.push(t)
-      }
-      // TODO deprecated
-      Object.assign(finalTags, {
-        then: () => finalTags,
-      })
-      return finalTags
-    },
     invalidate() {
       // Re-queue all current entries for normalization
       for (const entry of entries.values()) {
-        normalizeQueue.add(entry._i)
+        _normalizeQueue.add(entry._i)
       }
       head.dirty = true
       hooks.callHook('entries:updated', head)
