+++ packages/unhead/src/utils/index.ts
@@ -2,6 +2,8 @@ export * from './const'
 export { dedupeKey, hashTag, isMetaArrayDupeKey } from './dedupe'
 export { resolveMetaKeyType, resolveMetaKeyValue, resolvePackedMetaObjectValue, unpackMeta } from './meta'
 export { normalizeEntryToTags, normalizeProps } from './normalize'
+export { dedupeTags, resolveTags, resolveTitleTemplate, sanitizeTags } from './resolve'
+export type { ResolveTagsContext } from './resolve'
 export { sortTags, tagWeight } from './sort'
 export { processTemplateParams } from './templateParams'
 export { walkResolver } from './walkResolver'
