+++ packages/unhead/test/break-normalize-queue.test.ts
@@ -1,5 +1,6 @@
 import { describe, expect, it } from 'vitest'
 import { createUnhead } from '../src'
+import { resolveTags } from '../src/utils/resolve'
 
 describe('break Normalize Queue - Replicate Bug', () => {
   it('replicates the bug by following React pattern exactly', async () => {
@@ -9,7 +10,7 @@ describe('break Normalize Queue - Replicate Bug', () => {
     })
 
     // Step 1: Initial state - init values should be normalized
-    let tags = head.resolveTags()
+    let tags = resolveTags(head)
     expect(tags.find(t => t.tag === 'title')?.textContent).toBe('Init Title')
 
     // Step 2: Component mounts - React creates entry via push()
@@ -25,7 +26,7 @@ describe('break Normalize Queue - Replicate Bug', () => {
     })
 
     // Step 4: Resolve tags - component values should be active
-    tags = head.resolveTags()
+    tags = resolveTags(head)
     expect(tags.find(t => t.tag === 'title')?.textContent).toBe('Component Title')
 
     // Step 5: Component unmounts - React calls dispose()
@@ -36,7 +37,7 @@ describe('break Normalize Queue - Replicate Bug', () => {
     // Step 6: The bug manifests here
     // Without the fix, init entries won't be in normalizeQueue
     // so they won't be normalized and their _tags might be stale
-    tags = head.resolveTags()
+    tags = resolveTags(head)
 
     // This should pass but may fail with the original buggy code
     expect(tags.find(t => t.tag === 'title')?.textContent).toBe('Init Title')
@@ -52,7 +53,7 @@ describe('break Normalize Queue - Replicate Bug', () => {
     const initEntry = Array.from(head.entries.values())[0]
 
     // First resolution - init entry gets normalized
-    head.resolveTags()
+    resolveTags(head)
     const originalTags = initEntry._tags
     expect(originalTags).toBeDefined()
     expect(originalTags?.[0]?.textContent).toBe('Init Title')
@@ -61,7 +62,7 @@ describe('break Normalize Queue - Replicate Bug', () => {
     const componentEntry = head.push({ title: 'Component Title' })
 
     // Resolve - both entries get normalized
-    head.resolveTags()
+    resolveTags(head)
 
     // Dispose component entry
     componentEntry.dispose()
@@ -75,7 +76,7 @@ describe('break Normalize Queue - Replicate Bug', () => {
 
     // Force a scenario where init entry needs re-normalization
     // by modifying its input directly (simulating what might happen in complex scenarios)
-    const tags = head.resolveTags()
+    const tags = resolveTags(head)
 
     // If the bug exists, this might fail because init entry didn't get normalized
     expect(tags.find(t => t.tag === 'title')?.textContent).toBe('Init Title')
@@ -90,7 +91,7 @@ describe('break Normalize Queue - Replicate Bug', () => {
 
     // Step 1: Add component entry and resolve
     const componentEntry = head.push({ title: 'Component Title' })
-    head.resolveTags() // This processes and clears the normalize queue
+    resolveTags(head) // This processes and clears the normalize queue
 
     // Step 2: Manually trigger a state that requires re-normalization
     // In a real scenario, this might happen due to plugins or other factors
@@ -104,7 +105,7 @@ describe('break Normalize Queue - Replicate Bug', () => {
 
     // Step 4: The bug - if dispose() doesn't add remaining entries to normalizeQueue,
     // the init entry won't be normalized because its _tags is undefined
-    const tags = head.resolveTags()
+    const tags = resolveTags(head)
 
     // This should work but might fail if normalize queue bug exists
     expect(tags.find(t => t.tag === 'title')?.textContent).toBe('Init Title')
@@ -119,20 +120,20 @@ describe('break Normalize Queue - Replicate Bug', () => {
 
     // First mount
     const entry1 = head.push({ title: 'Component Title 1' })
-    head.resolveTags()
+    resolveTags(head)
 
     // First unmount (Strict Mode)
     entry1.dispose()
 
     // Second mount (Strict Mode remount)
     const entry2 = head.push({ title: 'Component Title 2' })
-    head.resolveTags()
+    resolveTags(head)
 
     // Second unmount (actual unmount)
     entry2.dispose()
 
     // Multiple disposals might expose the normalize queue bug
-    const tags = head.resolveTags()
+    const tags = resolveTags(head)
     expect(tags.find(t => t.tag === 'title')?.textContent).toBe('Init Title')
   })
 })
