+++ packages/unhead/test/unit/server/innerHTML.test.ts
@@ -1,6 +1,7 @@
 import type { ResolvableHead } from '../../../src/types'
 import { describe, it } from 'vitest'
 import { renderSSRHead } from '../../../src/server'
+import { resolveTags } from '../../../src/utils/resolve'
 import { createServerHeadWithContext } from '../../util'
 
 describe('ssr innerHTML', () => {
@@ -126,6 +127,6 @@ describe('ssr innerHTML', () => {
         },
       ],
     })
-    expect(head.resolveTags()).toMatchInlineSnapshot('[]')
+    expect(resolveTags(head)).toMatchInlineSnapshot('[]')
   })
 })
